mindlakevm/runtime/pyproject.toml
[project]
name = "mindlakevm-runtime"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.3",
    "openai==1.57.4",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "jsonschema==4.23.0",
    "tiktoken==0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
